fix: quality sweep addressing issues #78, #77, #61, #80, #46
- Split exercise_objectives.py (671 lines) into per-exercise modules
  under objectives/ package with core classes in objectives/__init__.py
  (issue #78)
- Refactor create_full_body() (253 lines) into 5 staged builder methods:
  _build_pelvis, _build_spine_and_head, _build_upper_limbs,
  _build_lower_limbs, _build_foot_contacts (issue #77)
- Remove redundant # type: ignore[import-not-found] on pydrake imports;
  mypy ignore_missing_imports=true already handles this (issue #61)
- Verify all public functions have return type annotations (100% coverage,
  mypy strict passes with zero errors) (issue #80)
- Add LOOP_CLOSURE.md documenting Drake barbell constraints, SDF tree
  invariant, single-parent attachment strategy, and runtime weld
  alternatives (issue #46)
