+++ src/drake_models/optimization/__init__.py
@@ -4,6 +4,8 @@ from drake_models.optimization.exercise_objectives import (
     BENCH_PRESS,
     CLEAN_AND_JERK,
     DEADLIFT,
+    GAIT,
+    SIT_TO_STAND,
     SNATCH,
     SQUAT,
     BalanceMode,
@@ -29,6 +31,8 @@ __all__ = [
     "DEADLIFT",
     "ExerciseObjective",
     "ExercisePhase",
+    "GAIT",
+    "SIT_TO_STAND",
     "SNATCH",
     "SQUAT",
     "TrajectoryConfig",
